% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruit.R
\docType{data}
\name{fruit}
\alias{fruit}
\title{Fruit Characteristics}
\format{
A data frame with 60 rows and 7 variables
}
\usage{
data(fruit)
}
\description{
A dataset containing the yield and other attributes of fruit
\itemize{
  \item Fruit Length
  \item Fruit Breadth
  \item Fruit Size
  \item Fruit Weight
  \item Fruit Volume
  \item Fruit Set
  \item Fruit Yield
}
}
\keyword{datasets}
